Clean manage routes, ADD about page

+++ src/app/blog/post/post.component.ts
@@ -21,8 +21,10 @@ export class PostComponent implements OnInit {
   }
 
   async delete(postId: string) {
-    await this._blog.deletePost(postId)
-    this._util.toastr.info('Post deleted')
+    const res = confirm('Are you sure?')
+    if (res) {
+      await this._blog.deletePost(postId)
+      this._util.toastr.info('Post deleted')
+    }
   }
-
 }
